move rescorer to trainingdata
