+++ meson.build
@@ -702,7 +702,7 @@ endif
 if get_option('rescorer')
   deps += subproject('gaviotatb').get_variable('gaviotatb_dep')
   executable('rescorer', 'src/rescorer_main.cc',
-       [common_files, 'src/rescorer/rescoreloop.cc'],
+       [common_files, 'src/trainingdata/rescoreloop.cc'],
        include_directories: includes, dependencies: deps, install: true)
 endif
